chg: update examples for POS client
