+++ examples/node/POS-client/approve-ERC20-for-deposit.js
@@ -1,15 +1,15 @@
 const bn = require('bn.js')
 
-const Matic = require('@maticnetwork/maticjs').default
-const config = require('./config')
-const utils = require('./utils')
+const MaticPOSClient = require('@maticnetwork/maticjs').MaticPOSClient
+const config = require('../config')
+const utils = require('../utils')
 
 const rootToken = config.ROPSTEN_TEST_TOKEN
 const amount = new bn(1).mul(utils.SCALING_FACTOR)
 const from = config.FROM_ADDRESS
 
 // Create Matic object
-const matic = new Matic({
+const maticPOSClient = new MaticPOSClient({
   maticProvider: config.MATIC_PROVIDER,
   parentProvider: config.PARENT_PROVIDER,
   rootChain: config.ROOTCHAIN_ADDRESS,
@@ -17,16 +17,15 @@ const matic = new Matic({
 })
 
 async function execute() {
-  matic.setWallet(config.PRIVATE_KEY)
-  return matic.approvePOSERC20ForDeposit(rootToken, amount, { from })
+  maticPOSClient.setWallet(config.PRIVATE_KEY)
+  return maticPOSClient.approveERC20ForDeposit(rootToken, amount, { from })
 }
 
 // eslint-disable-next-line
 async function executeRaw() {
-  await matic.initialize()
-  const txParams = await matic.approvePOSERC20ForDeposit(rootToken, amount, { from, encodeAbi: true })
+  const txParams = await maticPOSClient.approveERC20ForDeposit(rootToken, amount, { from, encodeAbi: true })
   let serializedTx = utils.buildRawTransaction(txParams, config.PRIVATE_KEY)
-  return matic.web3Client.parentWeb3.eth.sendSignedTransaction(serializedTx)
+  return maticPOSClient.web3Client.parentWeb3.eth.sendSignedTransaction(serializedTx)
 }
 
 // eslint-disable-next-line
