+++ examples/node/POS-client/deposit-ether-for-user.js
@@ -1,14 +1,14 @@
 const bn = require('bn.js')
 
-const Matic = require('@maticnetwork/maticjs').default
-const config = require('./config')
-const utils = require('./utils')
+const MaticPOSClient = require('@maticnetwork/maticjs').MaticPOSClient
+const config = require('../config')
+const utils = require('../utils')
 
 const amount = new bn(1).mul(utils.SCALING_FACTOR)
 const from = config.FROM_ADDRESS
 
 // Create Matic object
-const matic = new Matic({
+const maticPOSClient = new MaticPOSClient({
   maticProvider: config.MATIC_PROVIDER,
   parentProvider: config.PARENT_PROVIDER,
   rootChain: config.ROOTCHAIN_ADDRESS,
@@ -16,8 +16,8 @@ const matic = new Matic({
 })
 
 async function execute() {
-  matic.setWallet(config.PRIVATE_KEY)
-  return matic.depositPOSEtherForUser(from, amount, { from, gasPrice: '10000000000' })
+  maticPOSClient.setWallet(config.PRIVATE_KEY)
+  return maticPOSClient.depositEtherForUser(from, amount, { from, gasPrice: '10000000000' })
 }
 
 // eslint-disable-next-line
