+++ examples/node/POS-client/exit-ERC20.js
@@ -1,6 +1,6 @@
-const Matic = require('@maticnetwork/maticjs').default
+const MaticPOSClient = require('@maticnetwork/maticjs').MaticPOSClient
 
-const config = require('./config')
+const config = require('../config')
 
 const from = config.FROM_ADDRESS // from address
 
@@ -8,16 +8,16 @@ const from = config.FROM_ADDRESS // from address
 const burnTxHash = '0xbbaf7712bc426b665650992215f9c15ac9bda72bbb4e7d453ba9ed3875e0ebf5' // Insert txHash generated from burn-POS-ERC20.js
 
 // Create object of Matic
-const matic = new Matic({
+const maticPOSClient = new MaticPOSClient({
   maticProvider: config.MATIC_PROVIDER,
   parentProvider: config.PARENT_PROVIDER,
   rootChain: config.ROOTCHAIN_ADDRESS,
   posRootChainManager: config.POS_ROOT_CHAIN_MANAGER_ADDRESS,
 })
 
-matic.setWallet(config.PRIVATE_KEY)
-matic
-  .exitPOSERC20(burnTxHash, {
+maticPOSClient.setWallet(config.PRIVATE_KEY)
+maticPOSClient
+  .exitERC20(burnTxHash, {
     from,
   })
   .then(res => {
